Create search function for tag table
Most functionality requires us to search for all topics aka. leaf nodes
that relates to a tag, no matter how high in the tag tree the tag is.

Implementation is rudimentary and does not check for loops in the tag
table at the moment. These should be implemented and filtered out.

Scope: Tag Table