Add create DB tables script
All information will be stored in a local SQLite database. To ensure
consistency a basic script creates tables and basic level data.

The design only allows for ranges of levels for an item in the library.
This could pose a problem. Likewise it is not easy to differentiate what
type the reference is. A field should probably be added, if local
files starts appearing. So far the schema is only built with URLs in
mind.

Scope: Database