Create dummy UI
A simple UI is needed to interact with the database and find training
topic from the library. This implements the menu of a simple command
line UI.

The implementation is limited to a menu with only the exit option being
active. The command line nature of the UI restricts it to a local
system. Should probably be replaced by a web interface at some point.

Scope: UI