Create simple Web Item class
In order to assist the user in adding items, we create a web item class
for retrieving title and topic suggestions. This should speed up the
addition of new items.

The implementation is very rudimentary. There is no visual feedback for
the webpage being scraped. The topic suggestion can be very prone to
false positives.

Scope: Web