Create submenu for adding items to library
This submenu allows the user to add a new entry to the library. At the
same time it guides the user through relevant questions. The user can
store an item complete with topic and level ranges, or as To Read.

The back-end still needs to be created and we need to add a menu point
for getting a To Read article.

Scope: UI