Fix issue with quotes in titles
The current implementation has an issue with strings being terminated,
if the website title has a single quote in it.

To fix this all single and double quotes are replaced by space. This is
a crude and ugly solution, that should be improved when we have a
chance. Possibly it will be changed, when moving to SQLAlchemy.

Scope: Web