Add UI flow for read item in To Read
When the user adds an unread item to the list, it is entered with no
tag. This is the To Read logic. There needs to be a way for the user to
pull one item randomly from the stack.

Here we implement the UI flow of selecting a single item, opening it and
afterwards asking the user to give the item a tag.

Scope: UI