Retrieve and update library item logic
With the UI flow in place for picking a To Read item and asking for a
tag, we need to be able to produce an item and write the users updates
back into the database.

This is a simple function, that utilises existing definitions to create
a simple random read and an update function for the library table.

Scope: Database