Change tags table to map 1:N with items
The previous idea of having a graph like structure of my tags did not
give good results for the current setup. It was hard to settle on a
single topic for the items. In addition this made it difficult to
autogenerate tags.

Implementing a classic 1:N between item and tags creates a simpler
solution that can easily be populated.

This implementation limits the future possibilities of the project. It
will not be possible to traverse a search graph down through the
children. Perhaps we can implement this again at a later stage.

Scope: Database